Add Batch and BatchSearch objects
BatchSearch encapsulates the searching functionality
and exposes 2 methods for querying ET

Batch is an object for a specific batch from ET

Batch object has properties exposed for some of the relevant
data extracted it also exposes the raw data returned for the
batch as well as accessors methods to work with the raw returned data.

Batch can be constructed from data already partially broken down
from BatchSearch or to directly query ET and parse the data

the Batch::errata property will lazy load the Errata objects.
If you are just wanting to know which advisories you can look at
Batch::errata_ids property which will provide you with the list of
advisories ids that in the batch.